// src/types/Trade/ITradeResponse.ts
export interface ITrade {
  algCat: String;
  algId: String;
  avgPrc: String;
  brkClnt: String;
  cstFrm: String;
  exOrdId: String;
  exSeg: String;
  exTm: String;
  exp: String;
  flId: String;
  flLeg: number;
  flDt: String;
  fldQty: number;
  flTm: String;
  minQty: number;
  nReqId: String;
  nOrdNo: String;
  ordDur: String;
  prod: String;
  prcTp: String;
  prc: String;
  qty: number;
  rptTp: String;
  sym: String;
  tok: String;
  tm: String;
  trnsTp: String;
  trdSym: String;
  actId: String;
  brdLtQty: number;
  expDt: String;
  optTp: String;
  ordGenTp: String;
  posFlg: String;
  rmk: String;
  series: String;
  stat: String;
  stkPrc: String;
  usrId: String;
}